+++ JustFeedDesktop/bac.h
@@ -16,8 +16,8 @@ class Produit;
 class Bac
 {
   private:
-    Produit*     produit;     //!< produit que contient le bac
-    unsigned int poidsActuel; //!< à définir
+    Produit*     produit;                //!< produit que contient le bac
+    unsigned int poidsActuel;            //!< à définir
     double       pourcentageRemplissage; //!< pourcentage de remplissage du bu bac
 
   public:
@@ -35,7 +35,7 @@ class Bac
     // Mutateurs
     void setProduit(Produit* produit);
     void setNomProduit(const QString& nomProduit);
-    void setPrixProduit(const int& prixProduit);
+    void setPrixProduit(const double& prixProduit);
     void setPoidsActuel(int poidsActuel);
 };
